src/main/java/com/master/spring/spring2proj6/cdi/SomeCdiBusiness.java
package com.master.spring.spring2proj6.cdi;

import javax.inject.Inject;
import javax.inject.Named;

@Named
public class SomeCdiBusiness {

	@Inject
	private SomeCdiDao someCdiDao;

	public SomeCdiDao getSomeCdiDao() {
		return someCdiDao;
	}

	public void setSomeCdiDao(SomeCdiDao someCdiDao) {
		this.someCdiDao = someCdiDao;
	}

	public int findGreatest() {
		int greatest = Integer.MIN_VALUE;
		int[] data = someCdiDao.getData();
		for (int val : data) {
			if (val > greatest) {
				greatest = val;
			}
		}

		return greatest;
	}
}
